Update food&drink.html
In the first week, I constructed the framework of the web page and collected food&drink information, including the selection of food and restaurant, choose the corresponding introduction website. The second week, I added the corresponding images and hyperlinks. The third week, I adapted the code to make it viewable on computers and phones.